backend/requirements.txt
google-genai
uvicorn
dotenv
supabase
fastapi
pydantic